bot/requirements.txt
nonebot2~=2.0.0a14
nonebot-adapter-cqhttp~=2.0.0a14
nonebot-plugin-apscheduler==0.1.2
pydantic~=1.8.2
alchemical==0.5.0
superstream==0.1.1
jieba==0.42.1
numpy==1.21.4
peewee==3.14.8